+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 🐛 Bug fixes
 
-- fixed synapse recording indices to be within type (@kyralianaka)
+- fixed synapse recording indices to be within type (#643, @kyralianaka)
+- Fix inheriting from a Module #590 (#642, @jnsbck)
 
 # 0.9.0